refactor: break up large test file
