+++ tests/unit/test_fabric_run_pattern.py
@@ -2,1240 +2,47 @@
 
 This module tests the complete fabric_run_pattern tool functionality,
 including error handling, SSE response parsing, and API integration.
+
+The tests have been split across multiple files:
+- test_fabric_run_pattern_basic.py - Basic execution tests
+- test_fabric_run_pattern_error_handling.py - Error handling tests
+- test_fabric_run_pattern_streaming.py - Streaming functionality tests
+- test_fabric_run_pattern_validation.py - Input and parameter validation tests
+- test_fabric_run_pattern_advanced.py - Advanced features
+
+This file imports all the test classes to maintain backward compatibility.
 """
 
-from collections.abc import Callable
-from typing import Any
-from unittest.mock import patch
-
-import pytest
-import pytest_asyncio
-from fastmcp.tools import Tool
-from mcp import McpError
-
-from fabric_mcp.core import (
-    DEFAULT_MODEL,
-    DEFAULT_VENDOR,
-    FabricMCP,
-    PatternExecutionConfig,
+# Import all test classes to maintain backward compatibility
+from tests.unit.test_fabric_run_pattern_advanced import (
+    TestFabricRunPatternCoverageTargets,
+    TestFabricRunPatternModelInference,
+    TestFabricRunPatternRequestConstruction,
+    TestFabricRunPatternUnexpectedSSETypes,
+    TestFabricRunPatternVariablesAndAttachments,
 )
-from tests.shared.fabric_api.base import TestFixturesBase
-from tests.shared.fabric_api_mocks import (
-    FabricApiMockBuilder,
-    assert_mcp_error,
-    mock_fabric_api_client,
+from tests.unit.test_fabric_run_pattern_base import TestFabricRunPatternFixtureBase
+from tests.unit.test_fabric_run_pattern_basic import TestFabricRunPatternBasicExecution
+from tests.unit.test_fabric_run_pattern_error_handling import (
+    TestFabricRunPatternErrorHandling,
+)
+from tests.unit.test_fabric_run_pattern_streaming import TestFabricRunPatternStreaming
+from tests.unit.test_fabric_run_pattern_validation import (
+    TestFabricRunPatternInputValidation,
+    TestFabricRunPatternParameterValidation,
 )
 
-
-class TestFabricRunPatternFixtureBase(TestFixturesBase):
-    """Test cases for fabric_run_pattern tool SSE response handling."""
-
-    @pytest_asyncio.fixture
-    async def fabric_run_pattern_tool(
-        self, mcp_tools: dict[str, Tool]
-    ) -> Callable[..., Any]:
-        """Get the fabric_run_pattern tool from the server."""
-        # fabric_run_pattern is the 3rd tool (index 2)
-        return getattr(mcp_tools["fabric_run_pattern"], "fn")
-
-
-class TestFabricRunPatternBasicExecution(TestFabricRunPatternFixtureBase):
-    """Test cases for basic fabric_run_pattern tool execution scenarios."""
-
-    def test_successful_execution_with_basic_input(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test successful pattern execution with basic input."""
-        builder = FabricApiMockBuilder().with_successful_sse("Hello, World!")
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            result: dict[str, Any] = fabric_run_pattern_tool(
-                "test_pattern", "test input"
-            )
-
-            assert isinstance(result, dict)
-            assert "output_format" in result
-            assert "output_text" in result
-            assert result["output_text"] == "Hello, World!"
-            assert result["output_format"] == "text"
-            mock_api_client.close.assert_called_once()
-
-    def test_successful_execution_with_markdown_format(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test successful pattern execution with markdown output."""
-        builder = FabricApiMockBuilder().with_successful_sse(
-            "# Header\n\nContent", "markdown"
-        )
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            result: dict[str, Any] = fabric_run_pattern_tool(
-                "test_pattern", "test input"
-            )
-
-            assert result["output_text"] == "# Header\n\nContent"
-            assert result["output_format"] == "markdown"
-            mock_api_client.close.assert_called_once()
-
-    def test_successful_execution_with_complex_sse_response(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test pattern execution with complex SSE response containing
-        multiple chunks."""
-        sse_lines = [
-            'data: {"type": "content", "content": "First chunk", "format": "text"}',
-            'data: {"type": "content", "content": " Second chunk", "format": "text"}',
-            'data: {"type": "content", "content": " Final chunk", "format": "text"}',
-            'data: {"type": "complete"}',
-        ]
-        builder = FabricApiMockBuilder().with_sse_lines(sse_lines)
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            result: dict[str, Any] = fabric_run_pattern_tool(
-                "test_pattern", "test input"
-            )
-
-            assert result["output_text"] == "First chunk Second chunk Final chunk"
-            assert result["output_format"] == "text"
-            mock_api_client.close.assert_called_once()
-
-    def test_sse_response_with_empty_lines(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test SSE parsing to ensure handling of empty lines in the response."""
-        # Create SSE stream with empty lines mixed in
-        sse_lines = [
-            "",  # Empty line
-            'data: {"type": "content", "content": "Hello", "format": "text"}',
-            "",  # Another empty line
-            'data: {"type": "complete"}',
-            "",  # Final empty line
-        ]
-        builder = FabricApiMockBuilder().with_sse_lines(sse_lines)
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            result: dict[str, Any] = fabric_run_pattern_tool(
-                "test_pattern", "test input"
-            )
-
-            assert isinstance(result, dict)
-            assert result["output_text"] == "Hello"
-            assert result["output_format"] == "text"
-            mock_api_client.close.assert_called_once()
-
-
-class TestFabricRunPatternErrorHandling(TestFabricRunPatternFixtureBase):
-    """Test cases for fabric_run_pattern tool error handling."""
-
-    def test_network_connection_error(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test handling of network connection errors."""
-        builder = FabricApiMockBuilder().with_connection_error("Connection failed")
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            with pytest.raises(McpError) as exc_info:
-                fabric_run_pattern_tool("test_pattern", "test input")
-
-            # Connection errors should be wrapped in McpError by the tool wrapper
-            assert_mcp_error(exc_info, -32603, "Error executing pattern")
-            mock_api_client.close.assert_called_once()
-
-    def test_http_404_error(self, fabric_run_pattern_tool: Callable[..., Any]) -> None:
-        """Test handling of HTTP 404 errors."""
-        builder = FabricApiMockBuilder().with_http_error(404, "Pattern not found")
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            with pytest.raises(McpError) as exc_info:
-                fabric_run_pattern_tool("nonexistent_pattern", "test input")
-
-            assert_mcp_error(exc_info, -32603, "Fabric API returned error 404")
-            mock_api_client.close.assert_called_once()
-
-    def test_timeout_error(self, fabric_run_pattern_tool: Callable[..., Any]) -> None:
-        """Test handling of timeout errors."""
-        builder = FabricApiMockBuilder().with_timeout_error()
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            with pytest.raises(McpError) as exc_info:
-                fabric_run_pattern_tool("test_pattern", "test input")
-
-            assert_mcp_error(exc_info, -32603, "Request timed out")
-            mock_api_client.close.assert_called_once()
-
-    def test_sse_error_response(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test handling of SSE error responses."""
-        builder = FabricApiMockBuilder().with_sse_error("Pattern execution failed")
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            with pytest.raises(McpError) as exc_info:
-                fabric_run_pattern_tool("test_pattern", "test input")
-
-            assert_mcp_error(exc_info, -32603, "Pattern execution failed")
-            mock_api_client.close.assert_called_once()
-
-    def test_malformed_sse_data(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test handling of malformed SSE data."""
-        builder = FabricApiMockBuilder().with_partial_sse_data()
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            with pytest.raises(McpError) as exc_info:
-                fabric_run_pattern_tool("test_pattern", "test input")
-
-            assert_mcp_error(exc_info, -32603, "Malformed SSE data")
-            mock_api_client.close.assert_called_once()
-
-    def test_empty_sse_stream(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test handling of empty SSE stream."""
-        builder = FabricApiMockBuilder().with_empty_sse_stream()
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            with pytest.raises(McpError) as exc_info:
-                fabric_run_pattern_tool("test_pattern", "test input")
-
-            assert_mcp_error(exc_info, -32603, "Empty SSE stream")
-            mock_api_client.close.assert_called_once()
-
-    def test_sse_stream_with_non_data_lines(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test SSE stream processing with non-data lines (should be ignored)."""
-        sse_lines = [
-            ": This is a comment line",
-            'data: {"type": "content", "content": "Hello", "format": "text"}',
-            "event: test-event",
-            'data: {"type": "complete"}',
-            "",  # Empty line
-        ]
-        builder = FabricApiMockBuilder().with_sse_lines(sse_lines)
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            result: dict[str, Any] = fabric_run_pattern_tool(
-                "test_pattern", "test input"
-            )
-
-            assert result["output_text"] == "Hello"
-            assert result["output_format"] == "text"
-            mock_api_client.close.assert_called_once()
-
-
-class TestFabricRunPatternStreaming(TestFabricRunPatternFixtureBase):
-    """Test cases for fabric_run_pattern tool streaming functionality."""
-
-    def test_streaming_mode_with_simple_content(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test streaming mode returns generator with simple content."""
-        builder = FabricApiMockBuilder().with_successful_sse("Hello, World!")
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            result = fabric_run_pattern_tool("test_pattern", "test input", stream=True)
-
-            # Should return a generator for streaming mode
-            assert hasattr(result, "__iter__")
-
-            # Collect all chunks from the stream
-            chunks = list(result)
-
-            # Should have content and complete chunks
-            assert len(chunks) >= 1
-
-            # Find content chunks
-            content_chunks = [c for c in chunks if c.get("type") == "content"]
-            assert len(content_chunks) >= 1
-
-            # Verify chunk structure
-            content_chunk = content_chunks[0]
-            assert "type" in content_chunk
-            assert "format" in content_chunk
-            assert "content" in content_chunk
-            assert content_chunk["type"] == "content"
-            assert content_chunk["content"] == "Hello, World!"
-
-            mock_api_client.close.assert_called_once()
-
-    def test_streaming_mode_with_multiple_chunks(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test streaming mode with multiple content chunks."""
-        sse_lines = [
-            'data: {"type": "content", "content": "First", "format": "text"}',
-            'data: {"type": "content", "content": " Second", "format": "text"}',
-            'data: {"type": "content", "content": " Third", "format": "text"}',
-            'data: {"type": "complete"}',
-        ]
-        builder = FabricApiMockBuilder().with_sse_lines(sse_lines)
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            result = fabric_run_pattern_tool("test_pattern", "test input", stream=True)
-
-            chunks = list(result)
-            content_chunks = [c for c in chunks if c.get("type") == "content"]
-
-            # Should have 3 content chunks
-            assert len(content_chunks) == 3
-            assert content_chunks[0]["content"] == "First"
-            assert content_chunks[1]["content"] == " Second"
-            assert content_chunks[2]["content"] == " Third"
-
-            mock_api_client.close.assert_called_once()
-
-    def test_streaming_mode_with_different_formats(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test streaming mode handles different content formats."""
-        sse_lines = [
-            'data: {"type": "content", "content": "# Header", "format": "markdown"}',
-            'data: {"type": "content", "content": " Text", "format": "markdown"}',
-            'data: {"type": "complete", "format": "markdown"}',
-        ]
-        builder = FabricApiMockBuilder().with_sse_lines(sse_lines)
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            result = fabric_run_pattern_tool("test_pattern", "test input", stream=True)
-
-            chunks = list(result)
-            content_chunks = [c for c in chunks if c.get("type") == "content"]
-
-            # Verify format is preserved in streaming
-            assert len(content_chunks) == 2
-            assert content_chunks[0]["format"] == "markdown"
-            assert content_chunks[1]["format"] == "markdown"
-
-            mock_api_client.close.assert_called_once()
-
-    def test_streaming_mode_error_handling(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test streaming mode error handling during stream."""
-        builder = FabricApiMockBuilder().with_sse_error("Stream error occurred")
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            result = fabric_run_pattern_tool("test_pattern", "test input", stream=True)
-
-            # Should raise RuntimeError when iterating over the stream
-            with pytest.raises(RuntimeError) as exc_info:
-                list(result)
-
-            assert "Stream error occurred" in str(exc_info.value)
-            mock_api_client.close.assert_called_once()
-
-    def test_streaming_mode_malformed_sse_data(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test streaming mode with malformed SSE data."""
-        builder = FabricApiMockBuilder().with_partial_sse_data()
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            result = fabric_run_pattern_tool("test_pattern", "test input", stream=True)
-
-            # Should raise RuntimeError when processing malformed data
-            with pytest.raises(RuntimeError) as exc_info:
-                list(result)
-
-            assert "Malformed SSE data" in str(exc_info.value)
-            mock_api_client.close.assert_called_once()
-
-    def test_streaming_mode_empty_stream(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test streaming mode with empty SSE stream."""
-        builder = FabricApiMockBuilder().with_empty_sse_stream()
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            result = fabric_run_pattern_tool("test_pattern", "test input", stream=True)
-
-            # Should raise RuntimeError for empty stream
-            with pytest.raises(RuntimeError) as exc_info:
-                list(result)
-
-            assert "Empty SSE stream" in str(exc_info.value)
-            mock_api_client.close.assert_called_once()
-
-    def test_streaming_vs_non_streaming_behavior(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test that streaming vs non-streaming modes behave differently."""
-        sse_lines = [
-            'data: {"type": "content", "content": "First", "format": "text"}',
-            'data: {"type": "content", "content": " Second", "format": "text"}',
-            'data: {"type": "complete"}',
-        ]
-        builder = FabricApiMockBuilder().with_sse_lines(sse_lines)
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            # Test non-streaming mode (default)
-            non_streaming_result = fabric_run_pattern_tool(
-                "test_pattern", "test input", stream=False
-            )
-
-            # Should return dict with accumulated content
-            assert isinstance(non_streaming_result, dict)
-            assert "output_text" in non_streaming_result
-            assert "output_format" in non_streaming_result
-            assert non_streaming_result["output_text"] == "First Second"
-
-        # Reset the mock for streaming test
-        with mock_fabric_api_client(builder) as mock_api_client:
-            # Test streaming mode
-            streaming_result = fabric_run_pattern_tool(
-                "test_pattern", "test input", stream=True
-            )
-
-            # Should return generator
-            assert hasattr(streaming_result, "__iter__")
-            chunks = list(streaming_result)
-            content_chunks = [c for c in chunks if c.get("type") == "content"]
-
-            # Should have separate chunks, not accumulated
-            assert len(content_chunks) == 2
-            assert content_chunks[0]["content"] == "First"
-            assert content_chunks[1]["content"] == " Second"
-            # Note: close() is called twice in this test (once for non-streaming, once
-            # for streaming)
-            assert mock_api_client.close.call_count == 2
-
-
-class TestFabricRunPatternInputValidation(TestFabricRunPatternFixtureBase):
-    """Test cases for fabric_run_pattern tool input validation and edge cases."""
-
-    def test_empty_pattern_name_validation(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test that empty pattern name raises ValueError."""
-        # Test empty string
-        with pytest.raises(
-            ValueError, match="pattern_name is required and cannot be empty"
-        ):
-            fabric_run_pattern_tool("", "test input")
-
-        # Test whitespace-only string
-        with pytest.raises(
-            ValueError, match="pattern_name is required and cannot be empty"
-        ):
-            fabric_run_pattern_tool("   ", "test input")
-
-        # Test None (though this might be caught by type system)
-        with pytest.raises(
-            ValueError, match="pattern_name is required and cannot be empty"
-        ):
-            fabric_run_pattern_tool(None, "test input")
-
-    def test_empty_input_handling(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test pattern execution with empty input."""
-        builder = FabricApiMockBuilder().with_successful_sse("No input provided")
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            result: dict[str, Any] = fabric_run_pattern_tool("test_pattern", "")
-
-            assert result["output_text"] == "No input provided"
-            mock_api_client.close.assert_called_once()
-
-    def test_large_input_handling(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test pattern execution with large input."""
-        large_input = "x" * 10000
-        builder = FabricApiMockBuilder().with_successful_sse("Output")
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            result: dict[str, Any] = fabric_run_pattern_tool(
-                "test_pattern", large_input
-            )
-
-            assert result["output_text"] == "Output"
-            mock_api_client.close.assert_called_once()
-
-    def test_special_characters_in_pattern_name(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test pattern execution with special characters in pattern name."""
-        builder = FabricApiMockBuilder().with_successful_sse("Output")
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            result: dict[str, Any] = fabric_run_pattern_tool(
-                "pattern-with_special.chars", "test input"
-            )
-
-            assert result["output_text"] == "Output"
-            mock_api_client.close.assert_called_once()
-
-
-class TestFabricRunPatternParameterValidation(TestFabricRunPatternFixtureBase):
-    """Test cases for fabric_run_pattern parameter validation (Story 3.3)."""
-
-    def test_temperature_validation_valid_ranges(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test temperature parameter validation with valid ranges."""
-        builder = FabricApiMockBuilder().with_successful_sse("Valid temperature test")
-
-        with mock_fabric_api_client(builder):
-            # Test minimum valid temperature
-            result = fabric_run_pattern_tool(
-                "test_pattern", "test input", temperature=0.0
-            )
-            assert result["output_text"] == "Valid temperature test"
-
-            # Test maximum valid temperature
-            result = fabric_run_pattern_tool(
-                "test_pattern", "test input", temperature=2.0
-            )
-            assert result["output_text"] == "Valid temperature test"
-
-            # Test typical temperature
-            result = fabric_run_pattern_tool(
-                "test_pattern", "test input", temperature=0.7
-            )
-            assert result["output_text"] == "Valid temperature test"
-
-    def test_temperature_validation_invalid_ranges(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test temperature parameter validation with invalid ranges."""
-        # Test temperature too low
-        with pytest.raises(McpError) as exc_info:
-            fabric_run_pattern_tool("test_pattern", "test input", temperature=-0.1)
-        assert_mcp_error(
-            exc_info, -32602, "temperature must be a number between 0.0 and 2.0"
-        )
-
-        # Test temperature too high
-        with pytest.raises(McpError) as exc_info:
-            fabric_run_pattern_tool("test_pattern", "test input", temperature=2.1)
-        assert_mcp_error(
-            exc_info, -32602, "temperature must be a number between 0.0 and 2.0"
-        )
-
-        # Test invalid type
-        with pytest.raises(McpError) as exc_info:
-            fabric_run_pattern_tool("test_pattern", "test input", temperature="invalid")
-        assert_mcp_error(
-            exc_info, -32602, "temperature must be a number between 0.0 and 2.0"
-        )
-
-    def test_top_p_validation_valid_ranges(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test top_p parameter validation with valid ranges."""
-        builder = FabricApiMockBuilder().with_successful_sse("Valid top_p test")
-
-        with mock_fabric_api_client(builder):
-            # Test minimum valid top_p
-            result = fabric_run_pattern_tool("test_pattern", "test input", top_p=0.0)
-            assert result["output_text"] == "Valid top_p test"
-
-            # Test maximum valid top_p
-            result = fabric_run_pattern_tool("test_pattern", "test input", top_p=1.0)
-            assert result["output_text"] == "Valid top_p test"
-
-    def test_top_p_validation_invalid_ranges(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test top_p parameter validation with invalid ranges."""
-        # Test top_p too low
-        with pytest.raises(McpError) as exc_info:
-            fabric_run_pattern_tool("test_pattern", "test input", top_p=-0.1)
-        assert_mcp_error(exc_info, -32602, "top_p must be a number between 0.0 and 1.0")
-
-        # Test top_p too high
-        with pytest.raises(McpError) as exc_info:
-            fabric_run_pattern_tool("test_pattern", "test input", top_p=1.1)
-        assert_mcp_error(exc_info, -32602, "top_p must be a number between 0.0 and 1.0")
-
-    def test_penalty_validation_valid_ranges(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test penalty parameter validation with valid ranges."""
-        builder = FabricApiMockBuilder().with_successful_sse("Valid penalty test")
-
-        with mock_fabric_api_client(builder):
-            # Test presence_penalty valid range
-            result = fabric_run_pattern_tool(
-                "test_pattern", "test input", presence_penalty=-2.0
-            )
-            assert result["output_text"] == "Valid penalty test"
-
-            result = fabric_run_pattern_tool(
-                "test_pattern", "test input", presence_penalty=2.0
-            )
-            assert result["output_text"] == "Valid penalty test"
-
-            # Test frequency_penalty valid range
-            result = fabric_run_pattern_tool(
-                "test_pattern", "test input", frequency_penalty=-2.0
-            )
-            assert result["output_text"] == "Valid penalty test"
-
-            result = fabric_run_pattern_tool(
-                "test_pattern", "test input", frequency_penalty=2.0
-            )
-            assert result["output_text"] == "Valid penalty test"
-
-    def test_penalty_validation_invalid_ranges(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test penalty parameter validation with invalid ranges."""
-        # Test presence_penalty too low
-        with pytest.raises(McpError) as exc_info:
-            fabric_run_pattern_tool("test_pattern", "test input", presence_penalty=-2.1)
-        assert_mcp_error(
-            exc_info, -32602, "presence_penalty must be a number between -2.0 and 2.0"
-        )
-
-        # Test presence_penalty too high
-        with pytest.raises(McpError) as exc_info:
-            fabric_run_pattern_tool("test_pattern", "test input", presence_penalty=2.1)
-        assert_mcp_error(
-            exc_info, -32602, "presence_penalty must be a number between -2.0 and 2.0"
-        )
-
-        # Test frequency_penalty too low
-        with pytest.raises(McpError) as exc_info:
-            fabric_run_pattern_tool(
-                "test_pattern", "test input", frequency_penalty=-2.1
-            )
-        assert_mcp_error(
-            exc_info, -32602, "frequency_penalty must be a number between -2.0 and 2.0"
-        )
-
-        # Test frequency_penalty too high
-        with pytest.raises(McpError) as exc_info:
-            fabric_run_pattern_tool("test_pattern", "test input", frequency_penalty=2.1)
-        assert_mcp_error(
-            exc_info, -32602, "frequency_penalty must be a number between -2.0 and 2.0"
-        )
-
-    def test_model_name_validation(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test model_name parameter validation."""
-        builder = FabricApiMockBuilder().with_successful_sse("Valid model test")
-
-        with mock_fabric_api_client(builder):
-            # Test valid model names
-            result = fabric_run_pattern_tool(
-                "test_pattern", "test input", model_name="gpt-4"
-            )
-            assert result["output_text"] == "Valid model test"
-
-            result = fabric_run_pattern_tool(
-                "test_pattern", "test input", model_name="claude-3-opus"
-            )
-            assert result["output_text"] == "Valid model test"
-
-        # Test invalid model names
-        with pytest.raises(McpError) as exc_info:
-            fabric_run_pattern_tool("test_pattern", "test input", model_name="")
-        assert_mcp_error(exc_info, -32602, "model_name must be a non-empty string")
-
-        with pytest.raises(McpError) as exc_info:
-            fabric_run_pattern_tool("test_pattern", "test input", model_name="   ")
-        assert_mcp_error(exc_info, -32602, "model_name must be a non-empty string")
-
-    def test_strategy_name_validation(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test strategy_name parameter validation."""
-        builder = FabricApiMockBuilder().with_successful_sse("Valid strategy test")
-
-        with mock_fabric_api_client(builder):
-            # Test valid strategy names
-            result = fabric_run_pattern_tool(
-                "test_pattern", "test input", strategy_name="creative"
-            )
-            assert result["output_text"] == "Valid strategy test"
-
-            result = fabric_run_pattern_tool(
-                "test_pattern", "test input", strategy_name="analytical"
-            )
-            assert result["output_text"] == "Valid strategy test"
-
-        # Test invalid strategy names
-        with pytest.raises(McpError) as exc_info:
-            fabric_run_pattern_tool("test_pattern", "test input", strategy_name="")
-        assert_mcp_error(exc_info, -32602, "strategy_name must be a non-empty string")
-
-        with pytest.raises(McpError) as exc_info:
-            fabric_run_pattern_tool("test_pattern", "test input", strategy_name="   ")
-        assert_mcp_error(exc_info, -32602, "strategy_name must be a non-empty string")
-
-    def test_parameter_combinations(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test multiple parameter combinations work together."""
-        builder = FabricApiMockBuilder().with_successful_sse(
-            "Parameter combination test"
-        )
-
-        with mock_fabric_api_client(builder):
-            # Test all parameters together
-            result = fabric_run_pattern_tool(
-                "test_pattern",
-                "test input",
-                model_name="gpt-4",
-                temperature=0.8,
-                top_p=0.95,
-                presence_penalty=0.1,
-                frequency_penalty=-0.1,
-                strategy_name="creative",
-            )
-            assert result["output_text"] == "Parameter combination test"
-
-            # Test partial parameter combinations
-            result = fabric_run_pattern_tool(
-                "test_pattern",
-                "test input",
-                model_name="claude-3-opus",
-                temperature=0.3,
-                strategy_name="analytical",
-            )
-            assert result["output_text"] == "Parameter combination test"
-
-
-class TestFabricRunPatternRequestConstruction(TestFabricRunPatternFixtureBase):
-    """Test cases for fabric_run_pattern request construction with new parameters."""
-
-    def test_request_construction_with_new_parameters(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test that new parameters are correctly included in API request."""
-        builder = FabricApiMockBuilder().with_successful_sse(
-            "Request construction test"
-        )
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            fabric_run_pattern_tool(
-                "test_pattern",
-                "test input",
-                model_name="claude-3-opus",
-                temperature=0.8,
-                top_p=0.95,
-                presence_penalty=0.1,
-                frequency_penalty=-0.1,
-                strategy_name="creative",
-            )
-
-            # Verify API was called with correct parameters
-            mock_api_client.post.assert_called_once()
-            call_args = mock_api_client.post.call_args
-            payload = call_args[1]["json_data"]
-
-            # Check prompt-level parameters
-            prompt = payload["prompts"][0]
-            assert prompt["model"] == "claude-3-opus"
-            assert prompt["strategyName"] == "creative"
-
-            # Check root-level LLM parameters
-            assert payload["temperature"] == 0.8
-            assert payload["topP"] == 0.95
-            assert payload["presencePenalty"] == 0.1
-            assert payload["frequencyPenalty"] == -0.1
-
-    def test_request_construction_with_defaults(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test that omitted parameters use appropriate defaults."""
-        builder = FabricApiMockBuilder().with_successful_sse("Default test")
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            fabric_run_pattern_tool("test_pattern", "test input")
-
-            # Verify API was called with defaults
-            mock_api_client.post.assert_called_once()
-            call_args = mock_api_client.post.call_args
-            payload = call_args[1]["json_data"]
-
-            # Check default values are used
-            assert payload["temperature"] == 0.7  # Default temperature
-            assert payload["topP"] == 0.9  # Default top_p
-            assert payload["frequencyPenalty"] == 0.0  # Default frequency_penalty
-            assert payload["presencePenalty"] == 0.0  # Default presence_penalty
-
-            prompt = payload["prompts"][0]
-            assert prompt["strategyName"] == ""  # Default empty strategy
-
-    def test_backward_compatibility(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test that existing calls without new parameters still work."""
-        builder = FabricApiMockBuilder().with_successful_sse(
-            "Backward compatibility test"
-        )
-
-        with mock_fabric_api_client(builder):
-            # Test basic call (original Story 3.1 format)
-            result = fabric_run_pattern_tool("test_pattern", "test input")
-            assert result["output_text"] == "Backward compatibility test"
-            assert "output_format" in result
-
-            # Test with stream parameter (original Story 3.1 format)
-            result = fabric_run_pattern_tool("test_pattern", "test input", stream=False)
-            assert result["output_text"] == "Backward compatibility test"
-
-            # Test with config parameter (original Story 3.1 format)
-            config = PatternExecutionConfig()
-            result = fabric_run_pattern_tool(
-                "test_pattern", "test input", config=config
-            )
-            assert result["output_text"] == "Backward compatibility test"
-
-
-class TestFabricRunPatternModelInference(TestFabricRunPatternFixtureBase):
-    """Test cases for fabric_run_pattern tool model and vendor inference."""
-
-    @pytest.fixture
-    def server_no_defaults(self) -> FabricMCP:
-        """Create a FabricMCP server instance with no default model/vendor."""
-        with patch("fabric_mcp.core.get_default_model", return_value=(None, None)):
-            return FabricMCP()
-
-    @pytest.fixture
-    def server_claude_model(self) -> FabricMCP:
-        """Create a FabricMCP server instance with Claude model default."""
-        with patch(
-            "fabric_mcp.core.get_default_model",
-            return_value=("claude-3-opus", None),
-        ):
-            return FabricMCP()
-
-    @pytest.fixture
-    def server_gpt_model(self) -> FabricMCP:
-        """Create a FabricMCP server instance with GPT model default."""
-        with patch(
-            "fabric_mcp.core.get_default_model",
-            return_value=("gpt-3.5-turbo", None),
-        ):
-            return FabricMCP()
-
-    @pytest_asyncio.fixture
-    async def fabric_run_pattern_tool_no_defaults(
-        self, server_no_defaults: FabricMCP
-    ) -> Callable[..., Any]:
-        """Get the fabric_run_pattern tool from server with no defaults."""
-        tools = await server_no_defaults.get_tools()
-        return getattr(tools["fabric_run_pattern"], "fn")
-
-    @pytest_asyncio.fixture
-    async def fabric_run_pattern_tool_claude(
-        self, server_claude_model: FabricMCP
-    ) -> Callable[..., Any]:
-        """Get the fabric_run_pattern tool from server with Claude model."""
-        tools = await server_claude_model.get_tools()
-        return getattr(tools["fabric_run_pattern"], "fn")
-
-    @pytest_asyncio.fixture
-    async def fabric_run_pattern_tool_gpt(
-        self, server_gpt_model: FabricMCP
-    ) -> Callable[..., Any]:
-        """Get the fabric_run_pattern tool from server with GPT model."""
-        tools = await server_gpt_model.get_tools()
-        return getattr(tools["fabric_run_pattern"], "fn")
-
-    def test_pattern_not_found_500_error(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test handling of 500 error indicating pattern not found."""
-        builder = FabricApiMockBuilder().with_http_error(
-            500, "no such file or directory: pattern not found"
-        )
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            with pytest.raises(McpError) as exc_info:
-                fabric_run_pattern_tool("nonexistent_pattern", "test input")
-                mock_api_client.close.assert_called_once()
-
-            # Should be transformed to Invalid params error
-            assert exc_info.value.error.code == -32602
-            assert (
-                "Pattern 'nonexistent_pattern' not found"
-                in exc_info.value.error.message
-            )
-
-    def test_generic_500_error(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test handling of generic 500 error (not pattern not found)."""
-        builder = FabricApiMockBuilder().with_http_error(
-            500, "Database connection failed"
-        )
-
-        with mock_fabric_api_client(builder) as client:
-            with pytest.raises(McpError) as exc_info:
-                fabric_run_pattern_tool("test_pattern", "test input")
-
-            error = exc_info.value
-            assert error.error.code == -32603  # Internal error
-            assert "Database connection failed" in error.error.message
-            client.close.assert_called_once()
-
-    def test_vendor_inference_from_model_name(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test vendor inference when no default vendor is configured."""
-        builder = FabricApiMockBuilder().with_successful_sse("Inference test")
-
-        # Test Claude model inference
-        with patch("fabric_mcp.core.get_default_model", return_value=(None, None)):
-            with mock_fabric_api_client(builder):
-                config = PatternExecutionConfig(model_name="claude-3-sonnet")
-                result = fabric_run_pattern_tool(
-                    "test_pattern", "test input", config=config
-                )
-                assert result["output_text"] == "Inference test"
-
-        # Test GPT model inference
-        with patch("fabric_mcp.core.get_default_model", return_value=(None, None)):
-            with mock_fabric_api_client(builder):
-                config = PatternExecutionConfig(model_name="gpt-4")
-                result = fabric_run_pattern_tool(
-                    "test_pattern", "test input", config=config
-                )
-                assert result["output_text"] == "Inference test"
-
-    def test_hardcoded_model_fallback_when_no_defaults(
-        self, fabric_run_pattern_tool_no_defaults: Callable[..., Any]
-    ) -> None:
-        """Test fallback to hardcoded default model when no environment defaults."""
-        builder = FabricApiMockBuilder().with_successful_sse("Hello, World!")
-
-        # Mock get_default_model to return None for both model and vendor
-        with mock_fabric_api_client(builder) as mock_api_client:
-            result: dict[str, Any] = fabric_run_pattern_tool_no_defaults(
-                "test_pattern", "test input"
-            )
-
-            assert isinstance(result, dict)
-            # Check that API was called (would use hardcoded defaults)
-            mock_api_client.post.assert_called_once()
-            call_args = mock_api_client.post.call_args
-            payload = call_args[1]["json_data"]
-            assert payload["prompts"][0]["model"] == DEFAULT_MODEL
-            assert payload["prompts"][0]["vendor"] == DEFAULT_VENDOR
-
-    def test_vendor_inference_for_claude_models(
-        self, fabric_run_pattern_tool_claude: Callable[..., Any]
-    ) -> None:
-        """Test vendor inference from Claude model names."""
-        builder = FabricApiMockBuilder().with_successful_sse("Hello, World!")
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            result: dict[str, Any] = fabric_run_pattern_tool_claude(
-                "test_pattern", "test input"
-            )
-
-            assert isinstance(result, dict)
-            # Check that vendor was inferred as "anthropic" for Claude models
-            mock_api_client.post.assert_called_once()
-            call_args = mock_api_client.post.call_args
-            payload = call_args[1]["json_data"]
-            assert payload["prompts"][0]["model"] == "claude-3-opus"
-            assert payload["prompts"][0]["vendor"] == DEFAULT_VENDOR
-
-    def test_vendor_inference_for_gpt_models(
-        self, fabric_run_pattern_tool_gpt: Callable[..., Any]
-    ) -> None:
-        """Test vendor inference from GPT model names."""
-        builder = FabricApiMockBuilder().with_successful_sse("Hello, World!")
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            result: dict[str, Any] = fabric_run_pattern_tool_gpt(
-                "test_pattern", "test input"
-            )
-
-            assert isinstance(result, dict)
-            # Check that vendor was inferred as "openai" for GPT models
-            mock_api_client.post.assert_called_once()
-            call_args = mock_api_client.post.call_args
-            payload = call_args[1]["json_data"]
-            assert payload["prompts"][0]["model"] == "gpt-3.5-turbo"
-            assert payload["prompts"][0]["vendor"] == DEFAULT_VENDOR
-
-
-class TestFabricRunPatternVariablesAndAttachments(TestFabricRunPatternFixtureBase):
-    """Test cases for fabric_run_pattern tool variables and attachments parameters."""
-
-    def test_variables_parameter_basic(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test fabric_run_pattern with variables parameter."""
-        builder = FabricApiMockBuilder().with_successful_sse("Hello, World!")
-        variables = {"key1": "value1", "key2": "value2"}
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            result: dict[str, Any] = fabric_run_pattern_tool(
-                "test_pattern", "test input", variables=variables
-            )
-
-            assert isinstance(result, dict)
-            assert result["output_text"] == "Hello, World!"
-
-            # Verify variables were passed to the API
-            mock_api_client.post.assert_called_once()
-            call_args = mock_api_client.post.call_args
-            payload = call_args[1]["json_data"]
-            assert payload["prompts"][0]["variables"] == variables
-            mock_api_client.close.assert_called_once()
-
-    def test_attachments_parameter_basic(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test fabric_run_pattern with attachments parameter."""
-        builder = FabricApiMockBuilder().with_successful_sse("Hello, World!")
-        attachments = ["file1.txt", "file2.pdf"]
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            result: dict[str, Any] = fabric_run_pattern_tool(
-                "test_pattern", "test input", attachments=attachments
-            )
-
-            assert isinstance(result, dict)
-            assert result["output_text"] == "Hello, World!"
-
-            # Verify attachments were passed to the API
-            mock_api_client.post.assert_called_once()
-            call_args = mock_api_client.post.call_args
-            payload = call_args[1]["json_data"]
-            assert payload["prompts"][0]["attachments"] == attachments
-            mock_api_client.close.assert_called_once()
-
-    def test_variables_and_attachments_together(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test fabric_run_pattern with both variables and attachments."""
-        builder = FabricApiMockBuilder().with_successful_sse("Hello, World!")
-        variables = {"name": "test", "type": "example"}
-        attachments = ["doc.txt", "data.json"]
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            result: dict[str, Any] = fabric_run_pattern_tool(
-                "test_pattern",
-                "test input",
-                variables=variables,
-                attachments=attachments,
-            )
-
-            assert isinstance(result, dict)
-            assert result["output_text"] == "Hello, World!"
-
-            # Verify both parameters were passed to the API
-            mock_api_client.post.assert_called_once()
-            call_args = mock_api_client.post.call_args
-            payload = call_args[1]["json_data"]
-            assert payload["prompts"][0]["variables"] == variables
-            assert payload["prompts"][0]["attachments"] == attachments
-            mock_api_client.close.assert_called_once()
-
-    def test_variables_empty_dict(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test fabric_run_pattern with empty variables dict."""
-        builder = FabricApiMockBuilder().with_successful_sse("Hello, World!")
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            result: dict[str, Any] = fabric_run_pattern_tool(
-                "test_pattern", "test input", variables={}
-            )
-
-            assert isinstance(result, dict)
-            assert result["output_text"] == "Hello, World!"
-
-            # Verify empty variables dict was passed
-            mock_api_client.post.assert_called_once()
-            call_args = mock_api_client.post.call_args
-            payload = call_args[1]["json_data"]
-            assert payload["prompts"][0]["variables"] == {}
-            mock_api_client.close.assert_called_once()
-
-    def test_attachments_empty_list(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test fabric_run_pattern with empty attachments list."""
-        builder = FabricApiMockBuilder().with_successful_sse("Hello, World!")
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            result: dict[str, Any] = fabric_run_pattern_tool(
-                "test_pattern", "test input", attachments=[]
-            )
-
-            assert isinstance(result, dict)
-            assert result["output_text"] == "Hello, World!"
-
-            # Verify empty attachments list was passed
-            mock_api_client.post.assert_called_once()
-            call_args = mock_api_client.post.call_args
-            payload = call_args[1]["json_data"]
-            assert payload["prompts"][0]["attachments"] == []
-            mock_api_client.close.assert_called_once()
-
-    def test_variables_none_excluded_from_payload(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test that None variables are excluded from API payload."""
-        builder = FabricApiMockBuilder().with_successful_sse("Hello, World!")
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            result: dict[str, Any] = fabric_run_pattern_tool(
-                "test_pattern", "test input", variables=None
-            )
-
-            assert isinstance(result, dict)
-            assert result["output_text"] == "Hello, World!"
-
-            # Verify variables is not in the payload when None
-            mock_api_client.post.assert_called_once()
-            call_args = mock_api_client.post.call_args
-            payload = call_args[1]["json_data"]
-            assert "variables" not in payload["prompts"][0]
-            mock_api_client.close.assert_called_once()
-
-    def test_attachments_none_excluded_from_payload(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test that None attachments are excluded from API payload."""
-        builder = FabricApiMockBuilder().with_successful_sse("Hello, World!")
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            result: dict[str, Any] = fabric_run_pattern_tool(
-                "test_pattern", "test input", attachments=None
-            )
-
-            assert isinstance(result, dict)
-            assert result["output_text"] == "Hello, World!"
-
-            # Verify attachments is not in the payload when None
-            mock_api_client.post.assert_called_once()
-            call_args = mock_api_client.post.call_args
-            payload = call_args[1]["json_data"]
-            assert "attachments" not in payload["prompts"][0]
-            mock_api_client.close.assert_called_once()
-
-
-class TestFabricRunPatternUnexpectedSSETypes(TestFabricRunPatternFixtureBase):
-    """Test cases for fabric_run_pattern tool with unexpected SSE data types."""
-
-    def test_unexpected_sse_data_type(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test handling of unexpected SSE data type."""
-        # Create SSE stream with unexpected type
-        sse_lines = [
-            'data: {"type": "unexpected_type", "content": "Some content"}',
-        ]
-        builder = FabricApiMockBuilder().with_sse_lines(sse_lines)
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            # Let's see what actually happens - it might return successfully
-            # but with empty content due to no 'complete' signal
-            try:
-                result = fabric_run_pattern_tool("test_pattern", "test input")
-                # If it doesn't raise an error, we should get an empty result
-                assert result["output_text"] == ""
-                assert result["output_format"] == "text"
-            except McpError as e:
-                # If it does raise an error, verify it's the expected one
-                error_msg = str(e)
-                assert (
-                    "Empty SSE stream" in error_msg
-                    or "Unexpected SSE data type" in error_msg
-                )
-
-            mock_api_client.close.assert_called_once()
-
-    def test_unexpected_sse_data_type_with_unknown_field(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test handling of SSE data with missing type field."""
-        # Create SSE stream with missing type field
-        sse_lines = [
-            'data: {"content": "Some content"}',  # Missing "type" field
-        ]
-        builder = FabricApiMockBuilder().with_sse_lines(sse_lines)
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            # This should trigger the unexpected SSE data type handling
-            try:
-                result = fabric_run_pattern_tool("test_pattern", "test input")
-                # If it doesn't raise an error, we should get an empty result
-                assert result["output_text"] == ""
-                assert result["output_format"] == "text"
-            except McpError as e:
-                # If it does raise an error, verify it's the expected one
-                error_msg = str(e)
-                assert (
-                    "Empty SSE stream" in error_msg
-                    or "Unexpected SSE data type" in error_msg
-                )
-
-            mock_api_client.close.assert_called_once()
-
-    def test_sse_error_response_detailed(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test handling of SSE error response with detailed error message."""
-        builder = FabricApiMockBuilder().with_sse_error("Pattern validation failed")
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            with pytest.raises(McpError) as exc_info:
-                fabric_run_pattern_tool("test_pattern", "test input")
-
-            assert_mcp_error(exc_info, -32603, "Pattern validation failed")
-            mock_api_client.close.assert_called_once()
-
-    def test_unexpected_sse_data_type_forces_exception(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test to ensure unexpected SSE type definitely triggers exception."""
-        # Create SSE stream with ONLY unexpected type (no complete)
-        sse_lines = [
-            'data: {"type": "weird_unknown_type", "content": "test"}',
-        ]
-        builder = FabricApiMockBuilder().with_sse_lines(sse_lines)
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            # This should either raise an error or return empty result
-            try:
-                result = fabric_run_pattern_tool("test_pattern", "test input")
-                # If no exception, it should return empty content
-                assert result["output_text"] == ""
-                assert result["output_format"] == "text"
-            except McpError as e:
-                # If it raises, it should be about unexpected type or empty stream
-                error_msg = str(e)
-                assert (
-                    "Unexpected SSE data type" in error_msg
-                    or "Empty SSE stream" in error_msg
-                )
-
-            mock_api_client.close.assert_called_once()
-
-
-class TestFabricRunPatternCoverageTargets(TestFabricRunPatternFixtureBase):
-    """Test cases to target specific missing coverage lines."""
-
-    def test_default_config_creation_without_explicit_config(
-        self, fabric_run_pattern_tool: Callable[..., Any]
-    ) -> None:
-        """Test calling fabric_run_pattern without config to trigger defaults."""
-        builder = FabricApiMockBuilder().with_successful_sse("Coverage test")
-
-        with mock_fabric_api_client(builder) as mock_api_client:
-            # Call without config parameter to trigger default config creation
-            result: dict[str, Any] = fabric_run_pattern_tool(
-                "test_pattern",
-                "test input",
-                # No config parameter provided - should trigger line 494
-            )
-
-            assert isinstance(result, dict)
-            assert result["output_text"] == "Coverage test"
-            assert result["output_format"] == "text"
-            mock_api_client.close.assert_called_once()
+# Ensure all test classes are available for pytest discovery
+__all__ = [
+    "TestFabricRunPatternFixtureBase",
+    "TestFabricRunPatternBasicExecution",
+    "TestFabricRunPatternErrorHandling",
+    "TestFabricRunPatternStreaming",
+    "TestFabricRunPatternInputValidation",
+    "TestFabricRunPatternParameterValidation",
+    "TestFabricRunPatternRequestConstruction",
+    "TestFabricRunPatternModelInference",
+    "TestFabricRunPatternVariablesAndAttachments",
+    "TestFabricRunPatternUnexpectedSSETypes",
+    "TestFabricRunPatternCoverageTargets",
+]
